requirements.txt
aio-mc-rcon==3.2.2
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.3.6
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
python-dotenv==1.0.1
rcon==2.4.6
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
